requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
newspaper3k==0.2.8
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
python-dotenv==1.0.0
google-api-python-client==2.104.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
nltk==3.8.1
openai==1.3.0
